requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
playwright==1.48.0
Pillow==10.4.0
numpy
aiosqlite==0.20.0
aiofiles==24.1.0
python-multipart==0.0.9
jinja2==3.1.4
